Intermediary commit

+++ sfr-dmc-tomcat/download.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
 
-#groupId=$1
-#artifactId=$2
-#versionId=$3
-#artifactType=$4
+groupId=$1
+artifactId=$2
+versionId=$3
+artifactType=$4
+envId=$5
 #groupId=fr.sfr.bt.sh.dmc
 #artifactId=eDMC
 #versionId=2.7.1
@@ -13,21 +14,17 @@ GROUP_ID=$(echo "$groupId" |  tr '.' '/')
 ARTIFACT_ID=$artifactId
 VERSION_ID=$versionId
 ARTIFACT_TYPE=$artifactType
+ENV_ID=$envId
 
-
-
-eval export $(cat .env)
-echo $LOCAL_REPO
-echo $LOCAL_CONFIG_DIR
-echo $LOCAL_BIN_DIR
+. ./.env
 artifactFolder=$LOCAL_REPO/$GROUP_ID/$ARTIFACT_ID/$VERSION_ID
-echo artifactFolder=$artifactFolder
 fullLocalArtifactPath=$artifactFolder/$ARTIFACT_ID-$VERSION_ID.$ARTIFACT_TYPE
-echo fullLocalArtifactPath=$fullLocalArtifactPath
-fullRemoteArtifactPath=$REPO_URL/$GROUP_ID/$ARTIFACT_ID/$VERSION_ID/$ARTIFACT_ID-$VERSION_ID.$ARTIFACT_TYPE
-echo fullRemoteArtifactPath=$fullRemoteArtifactPath
-echo $LOCAL_BIN_DIR
-mkdir -p $artifactFolder
-curl -o $fullLocalArtifactPath $fullRemoteArtifactPath
-mkdir -p $LOCAL_BIN_DIR/$ARTIFACT_ID/
-cp $fullLocalArtifactPath $LOCAL_BIN_DIR/$ARTIFACT_ID/$ARTIFACT_ID.$ARTIFACT_TYPE
+fullRemoteArtifactPath=$REPO_URL/$GROUP_ID/$ARTIFACT_ID/$VERSION_ID/$ARTIFACT_ID-$VERSION_ID$ENV_ID.$ARTIFACT_TYPE
+echo Download called with params:  $groupId $artifactId $versionId $artifactType $envId
+if [ ! -f $fullLocalArtifactPath ]; then
+    echo Downloading $fullRemoteArtifactPath to $fullLocalArtifactPath
+    mkdir -p $artifactFolder
+    curl -o $fullLocalArtifactPath $fullRemoteArtifactPath
+    mkdir -p $LOCAL_BIN_DIR/$ARTIFACT_ID/
+    cp $fullLocalArtifactPath $LOCAL_BIN_DIR/$ARTIFACT_ID/$ARTIFACT_ID.$ARTIFACT_TYPE
+fi
